fixed logic for saving comments(left to add procedure for computing rating) and added basic structure for saving book

+++ src/main/kotlin/com/fon/knjizararest/entity/Comment.kt
@@ -9,7 +9,7 @@ import javax.persistence.*
 data class Comment(
         @Id @GeneratedValue(strategy = GenerationType.IDENTITY) @Column(name = "comment_id", nullable = false) val commentId: Long,
         @Column(name = "text", nullable = false) val text: String,
-        @Column(name = "rating", nullable = false) val rating:Float,
+        @Column(name = "rating", nullable = false) val rating:Int,
         @ManyToOne @JoinColumn(name = "user_id", nullable = false) @JsonManagedReference val user: User,
         @ManyToOne @JoinColumn(name = "book_id", nullable = false) @JsonBackReference val book: Book
 )
